+++ winlogbeat/docs/winlogbeat-options.asciidoc
@@ -376,6 +376,9 @@ winlogbeat.event_logs:
     include_xml: true
 --------------------------------------------------------------------------------
 
+* This can have a significant impact on performance that can vary depending
+on your system specs.
+
 [float]
 ==== `event_logs.tags`
 
@@ -461,24 +464,6 @@ stopped. *{vista_and_newer}*
 Setting `no_more_events` to `stop` is useful when reading from archived event
 log files where you want to read the whole file then exit. There's a complete
 example of how to read from an `.evtx` file in the <<reading-from-evtx,FAQ>>.
-
-[float]
-==== `event_logs.api`
-
-This selects the event log reader implementation that is used to read events
-from the Windows APIs. When the value is set to `wineventlog-raw` Winlogbeat will
-replace the default XML event log reader with a more performant implementation.
-*{vista_and_newer}*
-
-[source,yaml]
---------------------------------------------------------------------------------
-winlogbeat.event_logs:
-  - name: ForwardedEvents
-    api: wineventlog-raw
---------------------------------------------------------------------------------
-
-* If `include_xml` is `true` the performance will be the same as the default API,
-as performance improvements are lost when parsing the XML.
 
 
 [float]
